Fixed Merge Conflicts and moved Create Events and Events Form Fields into new Folders.
