+++ lib/features/event/presentation/event_form_fields.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_form_builder/flutter_form_builder.dart';
 import 'package:form_builder_validators/form_builder_validators.dart';
 
-import 'form-fields/field_padding.dart';
+import '../../group/presentation/form-fields/field_padding.dart';
 
 
 class EventNameField extends StatelessWidget{
